+++ public/index.html
@@ -32,14 +32,6 @@
     -->
     <link href="//fonts.googleapis.com/css2?family=Lora:ital@1&family=Montserrat:wght@400;700&family=Open+Sans:wght@300;600&display=swap" rel="stylesheet" />
     <title>Chicago Artist Guide</title>
-
-    <!-- Mailchimp stuff -->
-    <link href="//cdn-images.mailchimp.com/embedcode/classic-10_7_dtp.css" rel="stylesheet" type="text/css">
-    <style type="text/css">
-      #mc_embed_signup{background:#fff; clear:left; font:14px Helvetica,Arial,sans-serif; }
-      /* Add your own Mailchimp form style overrides in your site stylesheet or in this style block.
-        We recommend moving this block and the preceding CSS link to the HEAD of your HTML file. */
-    </style>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
